Multiplayer things honestly, but not really, but yes, cuz it all has to do with multiplayer connectivity and updating, mostly just been hashing out details of how player positions work, map region generation and how they paint on screen. That has been a huge struggle personally cuz i suck at maths and I think maybe that stems mostly from me trying to have my own idea of how map regions should be loaded in a 2D game(me and my bright ideas), I'll add a description of how it works in the readme, but currently I'm having a lot a of issues with it, it is(supposed to be) fully dynamic and can be fixed to accommodate the dynamic math that goes along with how it works(i say fixed because I converted my old client over to this system which runs on the server, so there's probably methods that shouldn't exist being called somewhere they shouldn't be called), all methods that pertain to current locations, and what region a player is in can be dynamically calculated, which means there does not have to be any cached id's, session to session, regions can be loaded dynamically from the client,(which means other players' updates, and locale id's and region information can be easily drawn from simple math instead of having to rely on large update packet, and boundries can be checked from the server and collision reports can maintain a 1 byte length, the client can make its own directional calculations of whether or not it can move in another direction, and whether or not the client is right or wrong, the server will accommodate a collision response, so it's impossible to cheat in that aspect i.e, forced noclip, as long as the server and client maintain perfect location sync. In a perfect world, as long as the locational updates are handled right, a lapse in these syncs or locational updates could be a potential flag for abusing our software, this would be extremely useful to detect people trying to break our client via desync, climbing or movement glitches, as a ruthless sandbox MMO, our admins will be completely lacking in empathy for hackers and they will face a swift mac address ban. The point of this game is to be ruthless, and thats how our staff will be
A large part of that dynamic region information, is that the fastest code in any language is maths, the server can handle equations faster than any redundant loop check, which could cycle thousands of times per second in a server with thousands of players online 

The biggest changes in this update are the map editor rewrite(which is not working at all) and the log rewrite, and full thread pooling update on TaskPool. It is now functioning how I originally envisioned, automatically pooling tasks into threads depending on workload, if a task takes many milliseconds to accomplish its task, it should share a thread with other tasks that don't take much time, like chat queues etc, theres still more to be done with it and im not even sure its functioning 100% proper but its doing it's job for now, better than it was since we are handling large loads per region potentially, but not always 

The next big change is I updated my Log.java to be fully static and i'm going to actually use it both client and server side now, one thing i wanted to add but didn't get to was trace logging so whatever static body uses the print method will report itself, so we know exactly where the message is coming from, I'm sure this is probably an easy thing with class inheritance, but I haven't messed with any type of reflection code in like 10 years so i'm not gonna bother with that right now.


I don't even know what the fuck im doing but i'm gonna try and build this out to be a fully scalable framework, not just for a 2D game but easily converted to a 3D game, which I think is kind of cool. 

I'd be suprised if anyone made it this far reading, but if you did, thanks for following my progress :) this is a hobby i plan to cash in on, while remaining open source. I'm not trying to welcome hackers or software abuse but I think people who are truly serious about exploiting the game, will definitely decompile the code anyway. This is my way of saying; "Here, have the code, see what you can do with it and post it in the forums :)"